Updated image responses

+++ app/Http/Controllers/ImageController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Middleware\BelongsToAccount;
 use App\Http\Middleware\ValidateImageRequest;
-use App\Models\Property;
+use App\Http\Middleware\ValidateJWT;
 use App\Models\PropertyImage;
 use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
 use Exception;
@@ -12,8 +13,11 @@ use Illuminate\Http\Request;
 
 class ImageController extends Controller
 {
+    // Constructor method for ImageController
     public function __construct()
     {
+        $this->middleware(ValidateJWT::class);
+        $this->middleware(BelongsToAccount::class);
         $this->middleware(ValidateImageRequest::class);
     }
 
@@ -25,8 +29,6 @@ class ImageController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
-        $property = Property::whereId(request('id'))->first();
-
         try {
             $uploadedFile = $request->file('file')->storeOnCloudinary('CleanCut/Property');
 
@@ -39,7 +41,8 @@ class ImageController extends Controller
             return response()->json([
                 "type" => "Successful request",
                 "message" => "Image added successfully to property",
-                "property" => $property->refresh(),
+                "secureURL" => $uploadedFile->getSecurePath(),
+                "images" => PropertyImage::wherePropertyId(request('id'))->get()->map(fn($image) => $image->url)
             ], 200);
         } catch (Exception $e) {
             return response()->json([
@@ -58,26 +61,31 @@ class ImageController extends Controller
      */
     public function destroy(Request $request): JsonResponse
     {
-        $property = Property::whereId(request('id'))->first();
+        $imageModel = PropertyImage::wherePropertyId($request->id)->whereUrl($request->url);
 
-        $imageModel = PropertyImage::wherePropertyId(request('id'))->whereUrl(request('url'));
+        if ($imageModel->exists() > 0) {
+            try {
+                Cloudinary::destroy($imageModel->first()->public_id);
 
-        try {
-            Cloudinary::destroy($imageModel->first()->public_id);
+                $imageModel->delete();
 
-            $imageModel->delete();
-
+                return response()->json([
+                    "type" => "Successful request",
+                    "message" => "Image deleted successfully from property",
+                    "images" => PropertyImage::wherePropertyId(request('id'))->get()->map(fn($image) => $image->url)
+                ], 200);
+            } catch (Exception $e) {
+                return response()->json([
+                    "type" => "Bad Request",
+                    "message" => "Image could not be deleted from property",
+                    "error" => $e->getMessage(),
+                ], 400);
+            }
+        } else {
             return response()->json([
-                "type" => "Successful request",
-                "message" => "Image deleted successfully from property",
-                "property" => $property->refresh()
-            ], 200);
-        } catch (Exception $e) {
-            return response()->json([
-                "type" => "Bad Request",
-                "message" => "Image could not be deleted from property",
-                "error" => $e->getMessage(),
-            ], 400);
+                "type" => "Not found",
+                "message" => "Image not found",
+            ], 404);
         }
     }
 }
